requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
Django==4.1.7
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
numpy==1.25.1
openpyxl==3.0.9
pandas==1.4.1
psutil==5.9.5
python-dateutil==2.8.2
python-dotenv==0.15.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.4